Move useGetObject to @mysten/core (#10658)
## Description 

Describe the changes or additions included in this PR.

Move `useGetObject` @mysten/core, Part of `Pre mainnet Eng hardening`
