+++ apps/core/src/index.ts
@@ -10,6 +10,7 @@ export * from './hooks/useGetRollingAverageApys';
 export * from './utils/formatAmount';
 export * from './utils/roundFloat';
 export * from './hooks/useGetTransferAmount';
+export * from './hooks/useGetObject';
 export * from './PostHogAnalyticsProvider';
 export * from './utils/formatPercentageDisplay';
 export * from './hooks/useGetSystemState';
